feat: add buy price field to product items and introduce price input component

+++ core/modules/product/model/product_item.model.go
@@ -23,7 +23,7 @@ type ProductItem struct {
 	SkuIMEI string `json:"sku_imei" gorm:"type:varchar(255);default:null"`
 
 	// Value
-	BuyPrice float64 `json:"buy_price" gorm:"default:0;not null"`
+	BuyPrice uint `json:"buy_price" gorm:"default:0;not null"`
 
 	// relations
 	Category ProductCategory `json:"category" gorm:"foreignKey:CategoryID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
